edited component BIRD, added variables like jump height and changed gravity to represent percentage/1.0--1.0, added y velocity to the bird but no x velocity because the pipes will be moving, try to replaced the Quad.draw() method, within the birds draw method with Quad.draw(this)
Added a overloaded method  to the QUADS class, it takes in the value of a static component and draws it on the screen. this also means youc an draw one of its childeen, (dynamiccomponent)

Added Methods to the STATIC COMPONENT class, getHeight(), getWidth(),getXTopLeft(), GetYTopLeft(), which returns the position of the top left corner of a static component

Added Methiods and variables to the DYNAMICCOMPNENT class,
Speed - how many pixels a compnent moves
Velocity(X/Y)-speed but with direction

getVelocityX,getVelocityY(),getSpeed(), setXVelocity, setYVelocity, setSpeed()

emptied out the PIPE component, only used for drawing pipes in a lowcation

added a gapPipe COMPONENT- it utilizes the pipe component to create 2 pipes, a top and bottom that creates a pipe with a random gap in it, there are variables in it to customize thoings like the height of the gap and more. it does not draw itself  and does not apply its own physics.

added component PIPE LOOP loop of gap pipes, there is an arraylist of gap pipes that loops over itself in the physics in draw section, int he physics sectionwhenever it goes past the left side of the screen it deletes itself from the array list and adds a new gap pipe, wheich will be pretty random at times, in the draw method it just draws the top and bottom pipe of the current gappipe in the loop.

edited MAINMENU scene, just replaced the pipe with a pipeloop